Added technologies intro, a separator between groups, and moved the files to under home.
