+++ src/app/modules/home/home.component.html
@@ -48,6 +48,7 @@
 		<a class="nav-anchor" id="technologies"> </a>
 		<h2 class="h1">Technologies</h2>
 		<div class="education-content">
+			<p>{{ content.technologies.intro }}</p>
 			<app-technologies></app-technologies>
 		</div>
 	</section>
